Remove header + relocate sentence underneath
Removed Changelog header and relocated sentence under Changelog header. Readers will already know that the document is a changelog, so it's unnecessary to have a header reiterating the file name. The text under it ("All notable changes...") is moved to the "What is a changelog" question because it is an explanation about the file, so it's more sensible to place the sentence together with other information about the file.